Engines/requirements.txt
###### Required packages for TIDeMEC Toolchain ######
pyyaml 
ruamel.yaml
pandas 
tabulate
mitrecve
jsonschema
git
splunk-sdk
carbon-black-cloud-sdk
toml